refactor: isolate sign up form in a component
